requirements.txt
google.generativeai~=0.8.3
Telethon~=1.37.0
python-dotenv~=1.0.1
pymongo~=4.10.1

google~=3.0.0
aiogram~=3.15.0